implement semi working internet query crew
can query the internet, just that the agents suck at it (the python
script scraped "list of datasets for machine-learning research" which
gave 0 useful info for the task of writing a markdown table,
but it kinda works. I tried adding another agent to the list but that
performed worse (or unexpectedly).