fix tools for database reformat (#122)
* fix code execution with mock and fix typos

apparently the user proxy agent had max_auto_reply set to 1
which made it never execute the given code from the code writer agent
now it runs the code and outputs it into the terminal

* add decent solution for terminating properly

also added a mock that doesn't currently work, but will fix later
could either add option for agents to create code that queries the
internet or add langchain tooling that can be given in the frontend

* wip solutions for crew to query internet

does not want to work, kinda did work sometimes but seems like the
environment needs to have a way to import python modules. got code
errors for import when it finally ran a script... very sad

* implement semi working internet query crew

can query the internet, just that the agents suck at it (the python
script scraped "list of datasets for machine-learning research" which
gave 0 useful info for the task of writing a markdown table,
but it kinda works. I tried adding another agent to the list but that
performed worse (or unexpectedly).

* add description to agents

idk if it works yet, will do more testing later and see if its worth

* fix adding description to agents

had key errors, just made scuffed solution for now

* start to add tools wip

* add tooling with reading and moving files

* add testfiles to gitignore

* format code

* add tool fix for database reformat