feat(auth): Change auth redirect and remove callback
This commit changes the redirect URL after authentication from
'auth/callback' to 'app'. The 'auth/callback' route is removed as it's
no longer needed. The logout process is simplified by removing error
handling. The 'code' query parameter is now checked in the layout
server route, and if present, it's used to exchange for a session.
The user data check in the layout component is removed.