feat: Update dependencies and improve authentication flow (#245)
* feat(SideNav): simplify logout button implementation

The logout button in the SideNav component has been simplified. The
previous implementation used a dynamic list for the bottom navigation,
which was overkill for a single static button. The new implementation
directly includes the logout button in the markup. This makes the code
easier to understand and maintain.

* work on fixing up supabase stuff again

* I really dislike poetry

* authentication is working!

* feat(auth): Change auth redirect and remove callback

This commit changes the redirect URL after authentication from
'auth/callback' to 'app'. The 'auth/callback' route is removed as it's
no longer needed. The logout process is simplified by removing error
handling. The 'code' query parameter is now checked in the layout
server route, and if present, it's used to exchange for a session.
The user data check in the layout component is removed.