feat(rest): change llm models to use azure openai keys (#259)
* feat: change to azure llms and add uvicorn runner

This commit updates the import paths in `__init__.py` to be absolute
instead of relative. It also adds a uvicorn runner at the end of the
file to allow for direct execution. Additionally, the `comment_bot.py`
file has been updated to use `AzureChatOpenAI` instead of `ChatOpenAI`.

* refactor: Replace ChatOpenAI with AzureChatOpenAI

This commit replaces the usage of `ChatOpenAI` with `AzureChatOpenAI` in
`relevance_bot.py`. The change is made in the `create_chain`,
`summarize_submission`, and `filter_with_questions` functions. This
refactor is part of an effort to transition from OpenAI to Azure for
language model processing.

* fix: Correct Azure model names in relevance_bot.py

This commit corrects the Azure model names used in `relevance_bot.py`.
The model name in the `create_chain` function's docstring and the
`calculate_relevance` function call has been changed from
"gpt-4-turbo-preview" to "gpt-4-turbo". The model names in the
`summarize_submission` and `filter_with_questions` functions have been
changed from "gpt-4-turbo" to "gpt-35-turbo".

* fix: Remove testing things and remove import

* fix: Remove imports 2: electric boogaloo

* fix: Accidentally removed a dot in an import